# openapi.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/user":{"post":{"operationId":"UserController_createUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"User successfully created"},"400":{"description":"Bad request - validation failed"}},"summary":"Create a new user","tags":["user"]},"get":{"operationId":"UserController_findAll","parameters":[],"responses":{"200":{"description":"List of all users"}},"summary":"Get all users","tags":["user"]}},"/user/{id}":{"get":{"operationId":"UserController_findById","parameters":[{"name":"id","required":true,"in":"path","description":"User ID","schema":{"type":"number"}}],"responses":{"200":{"description":"User found"},"404":{"description":"User not found"}},"summary":"Get user by ID","tags":["user"]},"patch":{"operationId":"UserController_update","parameters":[{"name":"id","required":true,"in":"path","description":"User ID","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"200":{"description":"User updated"},"404":{"description":"User not found"}},"summary":"Update user by ID","tags":["user"]},"delete":{"operationId":"UserController_delete","parameters":[{"name":"id","required":true,"in":"path","description":"User ID","schema":{"type":"number"}}],"responses":{"204":{"description":"User deleted successfully"},"404":{"description":"User not found"}},"summary":"Delete user by ID","tags":["user"]}}},"info":{"title":"LFT Backend API","description":"API documentation for LFT Backend","version":"1.0","contact":{}},"tags":[{"name":"user","description":"User management endpoints"}],"servers":[{"url":"http://localhost:8080","description":"Local development server"}],"components":{"schemas":{"CreateUserDto":{"type":"object","properties":{"firstName":{"type":"string","description":"First name of the user","example":"John"},"lastName":{"type":"string","description":"Last name of the user","example":"Doe"},"dateOfBirth":{"type":"string","description":"Date of birth in ISO format","example":"1990-01-15"},"nickName":{"type":"string","description":"Nickname or preferred name","example":"johnd"},"sex":{"type":"string","description":"Gender of the user","enum":["MALE","FEMALE"],"example":"MALE"}},"required":["firstName","lastName","dateOfBirth","sex"]}}}}